Python/Server/requirements.txt
coloredlogs==7.3
pyzmq==17.0.0b1
